clang-tidy, int16, Void--, ImageData++
- tidy.sh script, runs clang-tidy
- ozCore, ozEngine, ozFactory, common, matrix mostly cleaned of reasonable
  issues reported by clang-tidy
- (u)short -> (u)int16
- non-const reference params replaced by pointers
- ozCore
  * added int16, uint16 types, ushort removed
  * Void removed
  * String move made properly, no more memcpy on whole class
  * File::isFile() -> isRegular()
- ozFactory
  * ImageData made properly, with getters and setters & co.

+++ src/matrix/Bot.cc
@@ -487,7 +487,7 @@ void Bot::onUpdate()
       if (stamina < 0.0f) {
         life += stamina * DROWNING_RATIO;
 
-        if ((uint(timer.ticks) + uint(index * 1025)) % Timer::TICKS_PER_SEC == 0) {
+        if ((timer.ticks + uint64(index) * 1025u) % Timer::TICKS_PER_SEC == 0) {
           addEvent(EVENT_DAMAGE, 1.0f);
         }
       }
